fix: Introduce gorm.Model on Product struct and fix other codes as related it

+++ api/controllers/products/products_controller.go
@@ -37,7 +37,7 @@ func GetProduct(c *gin.Context) {
 		return
 	}
 
-	product, getErr := services.GetProduct(productID)
+	product, getErr := services.GetProduct(uint(productID))
 	if getErr != nil {
 		c.JSON(getErr.Status, getErr)
 		return
